requirements.txt
amqp==2.5.2
appnope==0.1.0
arrow==0.15.4
astroid==2.2.5
attrs==19.3.0
autopep8==1.4.3
backcall==0.1.0
beautifulsoup4==4.8.1
bibtexparser==1.1.0
billiard==3.6.2.0
bleach==3.1.0
boto==2.49.0
boto3==1.10.37
botocore==1.13.37
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.0.12
django-cors-headers==3.2.0
django-filter==2.1.0
django-rest-framework==0.1.0
django-simple-email-confirmation==0.70
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.14
entrypoints==0.3
flake8-django==0.0.4
future==0.18.2
GDAL==2.4.4
gensim==3.8.1
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.15
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
kafka-python==2.0.1
kiwisolver==1.1.0
kombu==4.6.7
lazy-object-proxy==1.3.1
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
nltk==3.4.5
notebook==6.0.2
numpy==1.14.5
pandas==0.23.3
pandocfilters==1.4.2
parso==0.5.1
pdfminer==20191125
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.1.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.4
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycryptodome==3.9.4
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.0
pytz==2019.3
PyYAML==3.13
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.22
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
smart-open==1.9.0
soupsieve==1.9.5
sqlparse==0.3.0
stop-words==2018.7.23
stripe==2.42.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.3
typed-ast==1.3.1
uritemplate==3.0.1
urllib3==1.24.1
vine==1.3.0
virtualenv==16.4.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.1
WTForms==2.2.1
zipp==0.6.0
